requirements.txt
tk
Pillow
qrcode
opencv-python
